KnowledgeBased/CloseDomain/requirement.txt
pdfminer.six
pymysql
pandas
openai